Reduce dependancies between states and conditions in dialog and combat
